Add Invoice, Refund, Token functional test classes
